+++ src/functions/convSize.js
@@ -3,12 +3,6 @@ const _fatalError = require("./fatalError");
 /**
  * Converts a size in bytes to bytes, kilobytes, megabytes and gigabytes.
  *
- * Usage:
- * ```js
- * _convertSize(100000); // Should return: { bytes: 100000, kilobytes: 100, megabytes: 0.1, gigabytes: 0 }
- * _convertSize(100000, 5); // Should return: { bytes: 100000, kilobytes: 100, megabytes: 0.1, gigabytes: 0.0001 }
- * ```
- *
  * @param {number} bytes The bytes to convert.
  * @param {number} decimals Optional; determines how many decimals to keep in each value. Defaults to `3`.
  * @returns An object containing the converted sizes in bytes, kilobytes, megabytes and gigabytes, each as a number.