+++ src/functions/init/startupChecks.js
@@ -1,11 +1,11 @@
 const os = require("os");
 
 // Get variables
-const { GLOBAL_NAME } = require("../variables/constants");
+const { GLOBAL_NAME } = require("../../variables/constants");
 
 // Get functions
 const _startupError = require("./startupError");
-const _fatalError = require("./fatalError");
+const _fatalError = require("../fatalError");
 
 const startupChecks = () => {
   try {