+++ CHANGELOG.md
@@ -9,13 +9,14 @@ This is the official BubbleOS changelog! All features will be recorded in _'grou
 - BubbleOS will no longer display the full intro after the first time it has been opened on the system, to reduce clutter.
 - Added the `-h`/`--help` flag to the BubbleOS executable, which displays all of the BubbleOS startup arguments and what they do.
 - Added the `--reset` flag to the BubbleOS executable, which resets the configuration file.
-- Internally added the `InfoMessages` and `ConfigManager` classes. The `manageConfig` function is now deprecated.
+- Internally added the `InfoMessages` and `ConfigManager` classes.
 
 ### Changed/Fixed Features
 
-- Fixed an issue where BubbleOS would crash if it could not save a command to history in the configuration file, or read the history. If it encounters an error reading the file, it will now automatically reset the file and not crash BubbleOS.
 - The "timebomb disabled" warning will now only display if the system time exceeds the expiry date.
+- Fixed an issue where BubbleOS would crash if it could not save a command to history in the configuration file, or read the history. If it encounters an error reading the file, it will now automatically reset the file and not crash BubbleOS.
 - Fixed an issue where BubbleOS would crash when encountering an inaccessible directory in `dirtree`.
+- Improved maintainability of the `index.js` file and moved other startup processes to their own dedicated file internally.
 
 ### Removed Features